API-334: move JobParameter to model in order to be persist by Doctrine
