+++ src/Akeneo/Bundle/BatchBundle/Job/DoctrineJobRepository.php
@@ -3,7 +3,7 @@
 namespace Akeneo\Bundle\BatchBundle\Job;
 
 use Akeneo\Component\Batch\Job\BatchStatus;
-use Akeneo\Component\Batch\Job\JobParameters;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Akeneo\Component\Batch\Job\JobRepositoryInterface;
 use Akeneo\Component\Batch\Model\JobExecution;
 use Akeneo\Component\Batch\Model\JobInstance;
