+++ src/Akeneo/Component/Batch/Job/JobParametersValidator.php
@@ -3,6 +3,7 @@
 namespace Akeneo\Component\Batch\Job;
 
 use Akeneo\Component\Batch\Job\JobParameters\ConstraintCollectionProviderRegistry;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
