+++ src/Akeneo/Component/Batch/Job/JobRepositoryInterface.php
@@ -4,6 +4,7 @@ namespace Akeneo\Component\Batch\Job;
 
 use Akeneo\Component\Batch\Model\JobExecution;
 use Akeneo\Component\Batch\Model\JobInstance;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Akeneo\Component\Batch\Model\StepExecution;
 
 /**
