+++ src/Akeneo/Component/Batch/Updater/JobInstanceUpdater.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Component\Batch\Updater;
 
-use Akeneo\Component\Batch\Job\JobParameters;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Akeneo\Component\Batch\Job\JobParametersFactory;
 use Akeneo\Component\Batch\Job\JobRegistry;
 use Akeneo\Component\Batch\Model\JobInstance;
