+++ src/Pim/Bundle/InstallerBundle/FixtureLoader/JobInstancesBuilder.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\InstallerBundle\FixtureLoader;
 
 use Akeneo\Component\Batch\Item\ItemProcessorInterface;
-use Akeneo\Component\Batch\Job\JobParameters;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Akeneo\Component\Batch\Model\JobExecution;
 use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Model\StepExecution;
