+++ src/Pim/Component/Connector/spec/Archiver/FileReaderArchiverSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Component\Connector\Archiver;
 
 use Akeneo\Component\Batch\Item\ItemReaderInterface;
 use Akeneo\Component\Batch\Job\Job;
-use Akeneo\Component\Batch\Job\JobParameters;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Akeneo\Component\Batch\Job\JobRegistry;
 use Akeneo\Component\Batch\Model\JobExecution;
 use Akeneo\Component\Batch\Model\JobInstance;
