+++ src/Pim/Component/Connector/spec/Reader/File/Yaml/ReaderSpec.php
@@ -2,7 +2,7 @@
 
 namespace spec\Pim\Component\Connector\Reader\File\Yaml;
 
-use Akeneo\Component\Batch\Job\JobParameters;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Akeneo\Component\Batch\Model\StepExecution;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Connector\ArrayConverter\ArrayConverterInterface;
