+++ src/Pim/Component/Connector/spec/Writer/File/Csv/WriterSpec.php
@@ -2,7 +2,7 @@
 
 namespace spec\Pim\Component\Connector\Writer\File\Csv;
 
-use Akeneo\Component\Batch\Job\JobParameters;
+use Akeneo\Component\Batch\Model\JobParameters;
 use Akeneo\Component\Batch\Model\JobExecution;
 use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Model\StepExecution;
